+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## v9.0.4
+* New Condition: Torghast Anima Power Count
 * #1811 - The group/target point of a group will now be preserved when moving a shrunk group via click-and-drag.
+### Bug Fixes
+* #1840 - Talents granted by Torghast powers are now correctly reflected by the Talent Learned condition.
+
 ## v9.0.3
 ### Bug Fixes
 * #1824 - Fix incorrect detection of Defensive mode in the Pet Attack Mode condition.